+++ src/main/java/wooteco/subway/line/dto/LinesResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dto;
 
 public class LinesResponse {
     private Long id;
