+++ src/main/java/wooteco/subway/station/controller/StationController.java
@@ -1,8 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.station.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.station.repository.StationDao;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.domain.Station;
 
 import java.net.URI;
 import java.util.List;
